requirements.txt
python-telegram-bot==20.8
APScheduler==3.10.4
SQLAlchemy==2.0.25
python-dotenv==1.0.0
Pillow==10.2.0
matplotlib==3.8.2